New section: about added
in about section again @media screen was used to make 2 column. In .section-center max-width was set to 1170px because even if screen gets bigger that section should not be more than 1170px. Overflow: hidden was added in .about-picture-container class so if we scale it, it would be scaled more than size of container.